[Hd, HdSt] Geometric shader fails on basisCurves.
A set of basis curves could have null geometric shaders if the repr and
surface shader changed. This caused a crash, null dereference, in the batch
drawing code that assumes that a draw item has a valid geometric shader.
We now ensure they all do at the _GetRepr call time in _Sync.

Cleaned up the _GetRepr code and make it more similar for meshes and
basisCurves. The setting of Geometric Shaders is now done more consistently
by ensuring there is always a geometric shader on a drawItem after _Sync.

The calls to SetGeometericShader are also guarded against null sets and the
logic that previously needed it is now using overwriting instead. The shared
pointer code handles the reclaiming of the old geometric shader if it was
not-null.

Removed code that duplicated the functionality and caused multiple iterations
through the drawItems per mesh and basisCurve.

(Internal change: 1785475)