Address the problem described in https://groups.google.com/d/msg/usd-interest/7SxzTgBVfv4/R7KTsTOmAAAJ
Invert the behavior with respect to single-sample Alembic properties:
by default, preserve them as single timeSamples.  Deprecate the
"Usd:timeSamples" metadata, and instead add "Usd:singleSampleAsDefault"
which we add upon converting a USD file to Alembic, and encounter an
attribute that has a Default but no timeSamples.

"Asset" alembic files will now convert to USD as "timeful", which seems
weird at Pixar where assets are all authored @Default, but there's
not likely to be any behavior change as a result, and getting proper
behavior for single-frame simulation files is much more important.

Fixed some bugs in camera treatment, discovered while addressing unit
test failures, and specialized handling of known-uniform properties, since
we prefer authoring only to default for those, and the primary change here
will often cause their abc samples to come in as a timeSample.

Finally, removed _ReadOrientation() from the "fallback type" processing in the
writer.  fallback type handles any abc schema we don't explicitly handle, and
also "pure overs" and typeless defs that got into the Alembic via converting a
usd file.  In none of those cases do we want to assume the data is leftHanded -
especially in the USD "over" case, we're almost certainly breaking the data in
doing so.  We should only manufacture orientation for the specific schemas we
know about for which it is relevant, i.e. Subd and Polymesh.

Add USD_API to UsdSchemaRegistry::_GetSpecId(), as inlined methods call it.

(Internal change: 1785504)
(Internal change: 1785644)