Remove UsdStage::Close
This method was originally meant to provide an
explicit way for Python consumers to release
resources used by a UsdStage. However, this
method allows a UsdStage to be put into an
invalid state that cannot be easily detected
with the current API, and would need to be
guarded against in client code if such an API
existed.

To avoid these issues, we're simply removing
the method and relying on UsdStage releasing
its resources when its refcount goes to 0,
both in C++ and Python (under CPython, we expect
that a Usd.Stage should immediately be destroyed
when its refcount drops to 0)

(Internal change: 1785530)