[UsdImaging] Move UsdImaging to a common Prim Tracking and Updating mechanism (part 2).
As we are adding more prim types to the system, we wanted to make Scene Delegate more generic, relying on the Prim Adapters in UsdImaging to do the prim type specific work.
In this second part of the change, the focus is delegating responsibility for prim insert/delete/dirty bit manipulation to the adapters, and building out the index proxy to support that.

Changes:
- Added UsdImagingPrimAdapter::MarkDirty, Mark{Feature}Dirty to handle dirty bits in a type-specific way.
- Added UsdImagingPrimAdapter::_RemovePrim to handle prim removal.
- Removed UsdImagingDelegate::_MarkRprimOrInstancerDirty, in favor of the adapter's version.
- Removed UsdImagingDelegate's_rprimPaths and _lightMap.
- Added UsdImagingIndexProxy::Insert{R,S,B}Prim, Remove{R,S,B}Prim, Mark{R,S,B}PrimDirty (plus instancer versions).

(Internal change: 1786548)