Adjust usdview heuristic to create overs even on component models. While testing for performance regressions, I discovered that the change we made to make variant switching with payloads more friendly (i.e. do not open the stage unloaded, first) had a rather drastic performance effect on large scenes.  For our Large Test set, it added 12 seconds onto an original stage-open time of 34s, and with the "process all models" change made here, that 12 seconds jumped to over 25.
So I also reinstituted the "unloaded stage processing" change in a different
form: we open a stage unloaded, and use it to extract model paths for
which we will create overs in a fresh layer, which we then use as the
session layer for a NEW stage opened as the user requested - since many
layers are already open, the overhead added by this approach is
significantly less.

In the Large Test set, total open-time is now 35s.  This has benefit for all
sets/shots, though we thankfully don't yet have anything as big as Large Test.
On an active show's slowest current set, usdview stage open time drops, with
this change, from 5.6s to 4.7.

(Internal change: 1790098)