[Hd+Backends] Remove Clean Prims after PreSync. A while ago, a fix was added to Hydra for repr initialization, which meant that repr initialization occurred more frequently (whenever the collection changed).
This fix added the PreSync step to Hydra's Render Index sync, which performs this repr initialization, before going to delegate.  It was often the case, that
the repr had already been initialized and the prims were otherwise clean.  However, this caused a performance regression as the scene delegate needed
to process a large number of prims.

This changelist contains two changes to address this.
- First, the Rprim's InitRepr() is now only called if the InitRepr dirty bit is set.  This avoid some unnecessary work of looking up the repr in prims.  Once InitRepr is called, the dirty bit is cleaned.
- Secondly, after presync, the dirty bits are checked and if the prim is clean, the prim is removed from the sync request vector.

Changes:
- Added InitRepr to all rprims initial dirty bit state.
- _PreSyncRPrims now tests InitRepr dirty bit before looping over all reprs.  The dirty bit is cleaned after the loop.
- _PreSyncRequestVector() after running _PreSyncRPrims now removes any prims from the vector that are marked clean.

(Internal change: 1790118)