[Hd] Optimize Hd_SortedIds::Remove(). Recent reports have indicated an increase in shutdown time in UsdView.  Profiling has shown the problem is removing SdfPath's from Hydra's sorted id vector.
Previously, the algorithm worked by using std::vector's erase() method.  This kept the vector sorted after the remove and finding the ids to remove could be done in O(log n) time.  However, this meant shifting SdfPaths, resulting in O(n) SdfPath copies (which meant an atomic increment and decrement).

This change, changes the removal to be a constant time removal, involving an element swap and popping the back of the vector.  However, this leaves the vector unsorted, so a O(n) search is needed.

Changes:
- No longer sort id if unsorted in the Remove() method.
- If sorted, continue to use the binary search else fallback to linear search.
- To remove an element swap with the last element and put back.
- Set the sorted count to the element just before the one that was removed (if not earlier).

(Internal change: 1790132)