[usdMaya] avoid a crash that can occur with nested assemblies and variant set selections
The current assembly node code does not really deal with variant set
selections made in Maya correctly. In particular, if a nested assembly is
brought in under a parent assembly because of a variant selection made on
the parent assembly in Maya, trying to set the nested assembly to its
'Expanded' representation could result in a crash. This is because the
assembly is re-opening the USD stage but is not re-applying the variant
selection that was made on the parent assembly.

This change merely avoids a crash. Further work is necessary to correctly
support variant set selections in all assembly representations.

Fixes #288

(Internal change: 1792331)