[usdMaya] set the USD stage's session layer as the edit target when opening stages for reading
With the exception of usdWriteJobCtx, all of the places where we open USD
stages are for reading only. We set the edit target of those stages to
the session layer so that variant set selections and other transient edits
do not accidentally end up in the root layers of the stages and pollute the
shared stage cache.

(Internal change: 1792334)