[usdImaging] Performance optimization for property changes. Performance testing has shown that toggling viability in usdview for the first time is much slower than subsequent toggles.
Investigation showed the difference was caused by resync notices the first time and update notices on subsequent toggles.

This problem was two-fold.  Firstly, Usd was notifying the UsdImaging Delegate that a Prim needed resync.  A prior change fixed
that issue, so that now the delegate gets the resync notice for a property not the prim.

The reason why the message is a resync is that the property is being inserted rather than just a value change.

However, at this time, the delegate doesn't really care whether a property is being inserted or updated.  Therefore, this change modifies the property resync handling to use the same code path as with the update notice.  Individual adapters have the option to force a prim resync based on the property.

Changes:
- Removed _ResyncProperty.
- Now calls _RefreshObject() instead.

(Internal change: 1792447)