tf: TfIsWritable accounts for whether a file's mount is read-only.
Use faccessat to determine writability rather than checking the permissions
returned from stat. This is a Linux only change. Windows and MacOSX
continue to use the stat based approach.

We'd prefer to not check ARCH_OS_* macros in Tf directly. A future
refactoring will clean this up.

Note that the existing unit tests pass without modification with this
change. Adding a test to handle this new case is difficult since it would
need to rely on knowledge of the Pixar mount setup or be otherwise able to
construct its own read-only mounts.

(Internal change: 1792777)