[Usdview] Fix prim search bug introcuced in change 1786580.
When the item selection has changed, we only need to clear out
the property tree as its regenerated on the fly, not the prim view.
This change also makes us consistent in the type of the search results.
Previously, it had been a list at times, and a deque at others.

(Internal change: 1793785)