[pxrUsdMayaGL] fix for Hydra-drawn geometry disappearing in VP 2.0 with shadows enabled
This change fixes an issue that would cause the geometry drawn by Hydra
(USD proxy shape nodes) to disappear when Viewport 2.0 is active and
shadows are enabled. In that case, Maya makes multiple render passes, one
for shadows and one for color, and we were previously emptying the batch
renderer's queue and allowing the draw data to be deleted after the shadow
pass, leaving nothing to be rendered during the main color render pass.

This change also includes some minor code cleanup, such as wrapping the
Hydra _renderIndex in a std::unique_ptr.

(Internal change: 1794089)