[hd/usdImaging] Adding Batch Remove Optimization to Hydra. There's been a recent performance regression in Hydra shutdown performance.  It turned out the cause was removing elements one at a time from SortedId list.  A previous optimization may individual removes faster, but still not back to the original performance numbers. To solve that problem, this change introduces a batch remove mechanism to Hydra's render index called RemoveSubtree().  This will delete all prims or all types under the given root that are owned by the given scene delegate.
The scene delegate check is necessary as it is possible to have nested delegates and we don't want to remove prims that are not owned by the delegate as that would be scene delegate and render index state would be out of sync for the nested delegate.

RemoveSubtree() removes all R, S and B prims belonging to the delegate as well as any Instancers, ExtCompuations and Tasks.

For R, S and B prims, the code first uses the PrimGather routines to get all prims belonging to a subtree, but rather than returning the ids as a list.  A new API was added, so that a range can be returned.  As this is a subtree operation, there should always be either 0 or 1 ranges.  So the new Api SubtreeAsRange() returns false if there are no prims in the subtree.

The code then loops over every prim in the range, using the prim info maps, testing if the prim should be deleted (because the delegate matches) and if so deletes the prim info entry.  If any prim is found not to be owned by the delegate, the prim range is split.  After looping over all the prims, the ranges are used to delete the prims from the sorted id list (starting in reverse to maintain indexes).  There's an additional optimization in HdSortedIds::RemoveRange() for if the range includes every prim.

Remove Subtree for other prim types (Tasks, Instancers, ExtComputations), is simpler as it simply loops over all prims testing each for if it matches the remove conditions.  Tasks have a little complexity as only SceneTasks are owned by a delegate.  The HdTask doesn't have a delegate.  Tasks should be moved to follow the same mechansim as other prims where the render index maintains the delegate association.

Changes:
- GetDelegate() on HdInstancer is now a public method.
- Added SubtreeAsRange() to HdPRimGather to return the subtree as a start and end index rather than copy the list of SdfPaths (an expensive operation due to the atomic increments).
- Added RemoveSubtree() batch delete operation to HdPrimTypeIndex.
- Added new RemoveSubtree() batch delete operation to HdRenderIndex.  Added prim type specific versions for Rprims, Instancers, ExtComputations and Tasks.
- Added RemoveRange() from HdSortedId.  This is significantly faster than removing individual ids.
- UsdImagingDelegate destructor now uses render index's RemoveSubtree() to removing all prims belonging to that delegate.  This does mean that adapters are no longer involved, this reverts to the behavior before the common prim handling went into UsdImagingDelegate.  Most adapters don't store state, so this is ok.  The exception is instance adapters, but as we a destroying the delegate anyway, we let the adapter's destructor handle this case.  Further work on the adapters would be needed to invoke the batch operation from elsewhere in the delegate.

(Internal change: 1794099)