[pxrUsdMayaGL] fix declaration order of Hydra render delegate and render index
The Hydra render index is constructed with and is dependent on the render
delegate. At destruction time, the render index uses the delegate to
destroy Hydra prims, so the delegate must be destructed *after* the render
index. We enforce that ordering by declaring the render delegate *before*
the render index, since class members are destructed in reverse declaration
order.

(Internal change: 1794350)