[pxrUsdMayaGL] ensure that one Hydra draw is executed for each Viewport 2.0 render pass
Viewport 2.0 may execute a render in multiple passes (shadow, color, etc.),
and Maya sends a notification when all rendering has finished. Instead of
emptying the render queue at a particular pass, we wait for this
notification from Maya to reset the batch renderer state.

The legacy viewport renders in a single pass and there is no such
notification sent by Maya, so we still reset the batch renderer state
after the first Hydra draw for the legacy viewport.

(Internal change: 1794534)