[Hd] Fixed dirtyness handling for the geometric shader in HdStMesh and HdStBasisCurves
CL 1785475 introduced a regression wherein the dirty bit handling for setting/updating the geometric shader was done _only_ for the draw items corresponding to the input repr.
The draw items corresponding to other reprs were neither updated nor reset. This led to shader compilation errors when a collection used another repr - for example, changing the complexity (refine level) and then triggering the picking code.

The fix resets the geometric shader for all draw items of all the reprs (to null), and sets the geometric shader for each draw item of the input repr when it is null.
We guarantee that each draw item of the input repr has a non-null geometric shaders before returning from _GetRepr. This happens in UpdateDrawItemGeometricShader.

Changes:
- Reset geometric shaders for all draw items of all reprs for the relevant dirty bits
- Set geometric shaders for all draw items of the input repr
- Removed null check from SetGeometricShader

(Internal change: 1794628)