[Hd] Fixed a regression as a result of the 'dont-sync-invisible-prims' optimization CL 1785823 attempted to add an optimization wherein we don't sync invisible prims whose visibility isn't dirty by not adding them to the RPrimSyncMap during HdRenderIndex::SyncAll. Unfortunately, this created a regression wherein repr initialization (via HdRprim::InitRepr) never happened and the assert in _GetRepr was triggered via HdRenderIndex::GetDrawItems (i.e., rprim sync was skipped, and we were bitten during the execute phase wherein the command buffer for the draw item couldn't be generated because its repr was never initialized)
Instead, we now do this optimization during HdRprim::CanSkipDirtyBitPropagationAndSync, wherein for an invisible prim, we check if the NewRepr and DirtyVisibility dirty bits aren't set. If that's the case, we set the dirty bits in the _sync request_ (note: _not_ the change tracker) to Clean.
This lets us avoid pulling any dirty data during delegate sync, and eventually updating the GPU buffers unnecessarily.

While the earlier fix hoped to avoid syncing invisible prims altogether, the current design requires some reworking to allow for that. We will still pay for syncing an invisible prim to initialize its repr. We only avoid time-varying updates to the invisible prim.

Changes:
- Moved NewRepr to the HdChangeTracker::RPrimDirtyBits enum
- Removed the 'dont-sync-invisible-prims' optimization in HdRenderIndex::SyncAll
- Added HdRprim::CanSkipDirtyBitPropagationAndSync to prevent syncing of dirty invisible prims post initialization
- Added a perf log token to use for testing and possibly logging

(Internal change: 1794786)