Rework helper for flattening time sample map
Modified _GetTimeSampleMap to use public API and
removed it from the UsdStage header. This makes
it possible to move the rest of the helper functions
for flattening out of UsdStage as well.

As a bonus, the new implementation is more efficient.
It also fixes a subtle bug where calling GetMetadata
on asset-valued attributes to retrieve time samples
or default values would return SdfAssetPaths without
the resolved paths filled in.

(Internal change: 1795520)