[Usdview] Fix bug in tab selection.
Before this change, the function which controlled updates to the various
components of the tabview held by usdview(right corner) was also incidentally calling into
updates on the property tree in the bottom left corner of the screen. This change
removes that behavior so the properties arent regenerated when switching tabs.

Proper updating of the attribute view is handled in higher level functions which listen to
object selection changing.

(Internal change: 1796644)