Add UsdProperty::FlattenTo
This method flattens a property to a specified prim or
property in the current edit target by writing resolved
metadata and values from the source property to the
target property spec. This is similar to what is done by
UsdStage::Flatten, but with additional handling for
writing out fallback values if needed and remapping
attribute connections/relationship targets to the
destination prim.

(Internal change: 1797003)