[Hd] Fix for Bool primvar data transfer to GPU.
Hydra's current primvar data transfer mechanism (VtBufferSource) assumes that the CPU memory layout matches the GPU.

With booleans, the input data in the VtValue is expected to be C++ type bool and there is an issue when sizeof(bool) != sizeof(GLSL bool).

Therefore, data must expanded before copying to the GPU.  To handle this, VtBufferSource converts the type of the VtValue to an int, so it has the right size.  However, the GL type remains GL_BOOL, so that codegen still works.

Changes:
- In VtBufferSource, if the VtValue contains a bool, convert to int and populate the type data appropriately.

(Internal change: 1797280)