[usdImaging] Add IsInstancerAdapter() flag/virtual method, as a way for instancers to opt into the nested-instancer adapter assignment rules.
Previously, code was calling ShouldCullChildren() to find out if an adapter was an instancer or not (and therefore, whether to override its adapter assignment in the UsdImagingDelegate).  Cards (upcoming change) uses the ShouldCullChildren() feature, but isn't an instancer, so needed a way to separate these two semantics.

As a bonus, all NI and PI rely on the return value from Populate() for name mangling.  This doesn't change behavior, but makes it so that name mangling only happens in one place (GprimAdapter::_ResolveCachePath).

(Internal change: 1797626)