[usdImaging] Add support for model hierarchy cards rendering, based on the model:drawMode attribute.
This change adds small hooks in the USD populate and resync behaviors to detect prims being drawn in alternate draw modes (such as cards).
Those prims are then assigned an instance of drawModeAdapter, and scene traversal is stopped for that subtree.

UsdImagingDrawModeAdapter inserts proxy geometry, a proxy shader, and potentially any bound textures to create a standin geometry
object for the subtree.

There's a minor refactor in gprimAdapter to export name-mangling functionality to adapters which need it, but which aren't gprims.

(Internal change: 1797754)
(Internal change: 1797790)