Fixed UsdGeomXformable::GetOrderedXformOps() to handle empty op lists.
Previously the method would fail to set resetsXformStack if there
was no op order, or if not op order was authored and the env setting
USD_READ_OLD_STYLE_TRANSFORM was on, or if there were no ops.  The
client had no way of knowing the value had not been set and some
clients defensively set the value to false before the call but not
all of them.

In addition, with this change the method will always raise a coding
error if the resetsXformStack pointer is NULL.  This is the
documented behavior but it would not issue the error in the same
cases it failed to set resetsXformStack.

(Internal change: 1801598)