Fixed areaLightGeometrySource attribute when reading Usd into Katana.
Change 1795290 added several geometry group children but neglected
to move the one existing child, areaLightGeometrySource, into that
group builder.  As a result that attribute would be set in the
geometry group then the entire geometry group would be replaced by
the contents of the builder, discarding the attribute.

(Internal change: 1801599)