[pxrUsdMayaGL] fix for end render notification handler being removed after scene update
This change fixes an issue where the Viewport 2.0 end render notification
handler was being removed on scene update. We add a notification handler
with a fixed name for the kEndRenderSemantic, so the handler is not tied to
a particular instance of the batch renderer. When the Reset() method was
called during scene update, we constructed a new batch renderer and then
destructed the old one, but removeNotification() was being called in the
destructor and ended up leaving no handler registered.

Instead, we simply always call addNotification() when constructing a batch
renderer instance and do not call removeNotification(), since Maya ensures
that only one handler is registered for a given name and semantic location.
That handler then invokes the end render function on the global batch
renderer instance, if any. The handler was also made a static member method
of UsdMayaGLBatchRenderer so that MayaRenderDidEnd() no longer needs to be
public.

(Internal change: 1801752)