Adding a new API schema for representing collections of objects in USD. This builds on the existing UsdGeomCollectionAPI schema (which will be deprecated shortly) and enhances it by allowing includes/excludes targeting similar to UsdLuxLinkingAPI.
In addition to includes and excludes, you can also specify an
expansion-rule for how the paths that are included in a collection must
be expanded to determine its members. The expansion-rule is encoded
as a token valued attribute with one of the following values:
* explicitOnly - only paths in the includes rel targets and not in the excludes
rel targets belong to the collection.
* expandPrims - all the prims at or below the includes rel-targets (and not
under the excludes rel-targets) belong to the collection.  Any property paths
included in the collection would, of course, also be honored. This is the default
behavior as it satisfies most use cases.
* expandPrimsAndProperties - like expandPrims, but also includes all properties
on all matched prims.  We're still not quite sure what the use cases are for this,
but it allows us to capture a whole lot of UsdObjects very concisely.

A collection can also include another collection by "including" its collection-
path. However, a collection is not allowed to exclude another collection. This
keeps the membership determining logic simple and efficient. It's also
consistent with other systems that encode collections of objects (like MaterialX).

(Internal change: 1802108)
(Internal change: 1802117)
(Internal change: 1802144)
(Internal change: 1802147)
(Internal change: 1802168)