[Usdview] Add new settings module for saving application state, and added unit tests.
NOTE: This change uses a new settings file, and old settings are not converted. Default settings will be loaded.

The old settings module used a plain dict which was serialized to a file using cPickle whenever a setting was changed. This made it very easy to use but very difficult to maintain properly. Settings were read and updated in several places and in some cases were used in place of actual properties. On top of this, the cPickle file format is not human-readable making debugging settings difficult.

The new settings module keeps track of multiple independent sources of state data and presents each with a dict of its own, and saves all states to a JSON file. It also provides facilities to do type checking and custom validation to allow for both safe and concise code. The new module only loads state during initialization, but can save at any time (right now Usdview only saves state when it is closed). When a save is invoked, each source of state data is asked to record its state then the collection of states are serialized. This means we don't need to do file I/O whenever we change a setting, only when we invoke a save. There is also a notion of versioning so when destructive changes are made to the state schema multiple versions of Usdview can still coexist without tampering with each other. Finally, JSON is a much more readable format which makes debugging significantly easier.

The old settings module is left in temporarily because it is still used by a plugin. Eventually, plugins will be given a way to save their state in the new settings module (this is part of the new plugin system).

Usdview's splitter positions relied on the old settings module, and had significant updates to work with the new one. Specifically, default panel sizes were added and the panel sizes are now cached in a member variable while in viewer mode rather than storing them in the old settings dict.

(Internal change: 1802377)
(Internal change: 1802460)