Initial GafferPx baking to Usd.
This change adds the ability to bake edits in a GafferPx to a Usd
file and to apply edits from a Usd file.  Edits come either from
Katana or from the baked Usd at any given time for any given
gaffer;  there's a switch to choose which.

There are several parts to this change.  First, usdBake now has a
method for baking a gaffer, BakeGaffer(), that takes a node with
at least two inputs (the sources before and after Katana edits)
and a path where the Usd file should be written.  This code is
based on the shader/material baking code and fits partly within
that framework but also does some things its own way.

Second, there are site ops for reading gaffer objects that synthesize
parameters expected by GafferPx but aren't necessarily written to
Usd.  For example 'info.gaffer.hasSize' on a RectLight.  These are
normally created by the light package but those don't run when
using the baked file.

Third, GafferPx gets a version bump and adds two nodes, a PxrUsdIn
and a Switch.  The latter chooses the baked or Katana edits.  The
former not only reads the baked Usd but is also the node we use
when baking.  As such it has two input ports added:  one takes the
source coming into the gaffer and the source with the gaffer's
Katana edits applied (what used to be the output of the gaffer).
The gaffer node gets a few new methods to get/set the bake path
and switch between Katana and baked Usd data.

Fourth, the Lights tab now has support for baking.  It adds a
tool button to the right of the current gaffer name.  This button
drops down a menu with options to bake the gaffer, use the baked
Usd or the Katana edits, and to show additional UI.  The additional
UI resembles the top of a LiveGroup, with the path to the baked
file and the resolved path.  It also has a button to bake and a
checkbox to choose baked or Katana edits;  these duplicate the
functions on the drop down menu.  This is not final UI.  The
expectation is that gaffers will be baked and checked in
automatically when the nearest enclosing live group is checked in.

Fifth, PxrUsdKatanaUsdInPluginRegistry and PxrUsdInOp needed
minor tweaks to fix bugs or add missing features.

Sixth, NodeChooser in SGVUtils/HelperWidgets didn't have much
flexibility regarding the style of items in the combo box.  It
was possible but inconvenient to subclass to modify the behavior.
Now it (and NodeIndicator) supports a delegate object.  If the
delegate is set and has a 'nodeIndicatorWillUpdateStyle' method
then that method will be called with a Style object that can be
modified to adjust the style.  The Lights tab adds such a method
and makes itself the delegate;  the method changes the color and
icon of baked gaffers.

Seventh, using the gaffer node's setUsingBaked() method not only
changes the value of the switch node but also (un)locks the root
package group and the edit package stack nodes.  Since all light
edits for a gaffer are under these nodes, locking them causes the
Katana parameter UI to disallow edits to everything under them.
This is important since a) the edits do not apply to the baked
data and b) we can't see the effect of those edits when using the
baked data.

Eighth, the Inspector tab is now baking aware.  It watches for
gafferPx_useBakedChanged events and in response will update the
node indicator icon and enable/disable editing of values.

Finally, this isn't complete.  Current issues include:
* Baked data should be read-only in the Viewer.
* There should be a flush Usd button to force re-reading Usd.
* Not baking animation.
* No variable switch to globally enable/disable using baked Usd.
* LightGroup isn't implemented yet.

(Internal change: 1802390)