[pxrUsdMayaGL] fixes to address floating point error with lights in the legacy viewport
When using the MDrawRequest's matrix and modelViewMatrix along with
MMatrix::inverse() to compute the world-to-view matrix for Hydra, very
small bits of floating error were sometimes being introduced. That error
could result in the positions of lights being computed downstream as having
w coordinate values very close to but not exactly 1.0 or 0.0. When drawn,
the light would then flip between being a directional light (w = 0.0) and a
non-directional light (w = 1.0). Selection could sometimes affect whether
or not this error was present since it could affect the ordering of Draw()
calls from Maya, and the batch renderer only invokes Hydra to actually do
the drawing on the first Draw() call.

Instead, we get the inclusiveMatrix of the M3dView's camera, and then use
GfMatrix4d::GetInverse() to invert it. We then pass GfMatrix4d objects down
to the SetCameraState(), SetLightingStateFromVP1() and _RenderBatches()
methods instead of MMatrix objects. The parameter naming for the "view"
matrix is also clarified to indicate that the world-to-view matrix is
expected.

SetLightingStateFromVP1() now also takes the projectionMatrix as a
parameter in addition to the worldToViewMatrix so that it can use both to
set the camera on the GlfSimpleLightingContext.

Also, the task params for HdxSimpleLightTask include the viewport, so this
change includes updating/invalidating those task params when the viewport
changes.

(Internal change: 1802392)