Edit target mapping for composition arc authoring
The USD composition arc authoring APIs will now
map any non-root prim paths they are given to
the namespace of the current edit target.

This only occurs when the given paths are
expected to be in the stage's namespace. For
UsdInherits and UsdSpecializes, this is always the
case. For UsdReferences, this only occurs for
internal references.

(Internal change: 1802518)