[UsdImaging] Remove Path translation cache.
The UsdImaging delegate contained a path translation cache between Usd path and Hydra Render Index path.
The difference between these two paths is that the render index path includes the UsdImaging delegate's id as a prefix.

The translation cache was added to overcome the overhead of the path operations to modify the prefix.  However, as we threaded the code, it meant adding locks, which undone the benefits.

Removing the cache means we can lower the memory usage and also get some speedup from increased parallelism on Hydra (as the locks were shown to cause a single threaded bottleneck)

Changes:
- Removed _indexToUsdPathMap and _usdToIndexPathMap

- Removed associated RW mutexes.

- Now path translation functions just use the cache "missed" path as their implementation.

(Internal change: 1802575)