[Hd] Issue error when shaders fail to compile.
Recent efforts on improving internal testing results showed that a number of tests with legacy shaders were failing to compile, but went unnoticed as Hydra was silently falling back to the "fallback surface shader".

This change, makes shader compilation failure not so silient and issues a TF_CODING_ERROR if the shader fails to compile.

Changes:
- Added a TF_CODING_ERROR to drawBatch if the shader failed to compile.

(Internal change: 1802576)