[Usdview] Renamed MainWindow to AppController and removed QMainWindow superclass. AppController now owns a QMainWindow widget which itself contains no app data or logic. This was done because the main window needs to be exposed to plugins so they can create child windows, but we don't want to also give them access to internal app data and functionality.
Since AppController still contains a large amount of controller logic as well as the entire data model, AppController is still called by several of its subcomponents and plugins. This will be corrected when a separate data model and fully-featured plugin API are available.

Note that AppController inherits from QtCore.QObject because it acts as a Qt eventFilter for mainWindow (where MainWindow used to be its own eventFilter).

(Internal change: 1802759)