[hdx] sanity check HdxShadowParam's shadowMatrix computation pointer before executing it
This prevents a crash if a client fails to provide a computation pointer
in the shadow params.

Also, the extraction of the windowPolicy is pulled out of the light loop,
since that only needs to be fetched once.

Finally, if either the windowPolicy or the shadowMatrix are missing, we
simply disable shadows for that light and continue rather than returning
from _Sync(). This ensures that we don't leave the lighting context in a
half-configured state.

(Internal change: 1802900)