[usdMaya] fix for exporting over the same file that was imported
If we're exporting over a file that was previously imported, there may
still be stages in the stage cache that have that file as a root layer.
Creating a new stage with that file will fail because the layer already
exists in the layer registry, so we try to clear the layer from the
registry by erasing any stages in the stage cache with that root layer.

(Internal change: 1803019)