[Hd] Fix for fallback surface shader when a custom displacement shader is defined.
When shaders fail to compile, Hydra assumes the reason for the failure was the "surface" (material) shader and tries the compilation again using a fallback "surface" (material) shader.

The "surface" (material) shader, may or may not define the displacement shader terminal. If it is not defined the geometric shader will either define its own displacement shader terminal or if the whole geometry shader matches fixed-function, it disables the geometry shader completely.

Some internal tests were failing because the test uses a legacy surface shader that hasn't been updated yet and so was failing to compile and thus uses the fallback shader. However, some of these tests defined a custom displacement shader terminal, so the fallback case was also failing to compile because the displacement shader terminal was missing from the fallback shader.

However, if the fallback shader defined this terminal, the inverse set of tests were failing because the displacement shader was defined twice.

The current quick-fix solution is to handle this like a multiple include issue, as use include guards to make sure the displacement shader terminal's code fragment only gets added to the source once.

Changes:
- fallback surface shader now includes the displacement shader terminal from terminals.glslfx.
- Added include guards around the displacement shader terminal.

(Internal change: 1803177)