[UsdImaging] Moved GLSL material processing to a separate adapter
In preparation for more general processing of UsdShade material
bindings, this change moves the support for processing of GLSL
material networks to UsdImagingHydraMaterialAdapter.

This also makes the lookup of material adapters more
consistent with other kinds of prim adapters. This allows us
to remove the special one-off shader adapter owned by
the scene delegate, but requires us to add a way to skip
early population of the prims managed within a material
network.

For now, we redirect adapter look up for all Shader prims to
the legacy HydraPbsSurface material adapter. This is an
unfortunate temporary measure until we add an adapter
for general Shader network processing.

There is still some GLSL material specific logic remaining in
the gprimAdapter which does discovery of primvars and that will
be addressed separately.

(Internal change: 1803182)
(Internal change: 1803211)